Writing Merge Function for AppStore Deployement RestAPI BDRE-208

+++ md-dao/src/test/java/com/wipro/ats/bdre/md/dao/AppDeploymentQueueDAOTest.java
@@ -60,7 +60,7 @@ public class AppDeploymentQueueDAOTest {
         LOGGER.info("Adq is added with Id:" + adqId);
 
         appDeploymentQueueDAO.update(adq);
-        adq = appDeploymentQueueDAO.get(adqId.intValue());
+        adq = appDeploymentQueueDAO.get(adqId);
         assertEquals("Test Updated",adq.getAppName());
         LOGGER.info("Updated Description is:" + adqStatus.getDescription());
         appDeploymentQueueDAO.delete(adqId);
